SPP1-213 Move MinimalDataSet to its own module
This facilitates a move of the `dataset` fixture to `conftest.py` from
where it can be used by all unit tests and not just `test_dataset.py`.
